refactor(usecase): 유즈케이스 삭제
