+++ app/src/main/java/woowacourse/app/presentation/ui/main/home/HomeContract.kt
@@ -1,7 +1,6 @@
 package woowacourse.app.presentation.ui.main.home
 
 import woowacourse.app.presentation.model.HomeData
-import woowacourse.app.presentation.usecase.main.MainUseCase
 
 interface HomeContract {
     interface View {
@@ -10,7 +9,6 @@ interface HomeContract {
     }
 
     interface Presenter {
-        val mainUseCase: MainUseCase
         fun fetchHomeData()
     }
 }
